Refactored user context
